% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Utility functions}
\usage{
check_columns(data, required_columns)
}
\description{
Utility functions
}
\section{Functions}{

\describe{
\item{check_columns}{Checks if all required columns are present in the input data.}
\item{melt_data}{Transfers the data to a long format}
}
}

