% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_PFS.R
\name{correlate_PFS}
\alias{correlate_PFS}
\title{Calculate the correlation between PFS1 and PFS2}
\usage{
correlate_PFS(data, delta = 1)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) )}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}
}
\value{
Statistics of the correlation between PFS1 and PFS2
}
\description{
Calculate the correlation between PFS1 and PFS2
}
\examples{
data(input)
correlate_PFS(data = input, delta = 1.3)
}
