% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_PFSr.R
\name{count_PFSr}
\alias{count_PFSr}
\title{Analysis of progression free survival ratio (PFSr) using the count-based method}
\usage{
count_PFSr(data, delta = 1, prob = 0.5)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric)}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}

\item{prob}{a numeric value. The probability}
}
\value{
Count-based statistics of PFSr analysis
}
\description{
Analysis of progression free survival ratio (PFSr) using the count-based method
}
\examples{
data(input)
count_PFSr(data = data)
}
