% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSr_samplesize_ges.R
\name{ges_PFSr_samplesize}
\alias{ges_PFSr_samplesize}
\title{Study sample size calculation using the generalized effect size method}
\usage{
ges_PFSr_samplesize(
  ges = 0.3,
  alpha = 0.05,
  power = 0.8,
  lost = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{ges}{a numeric value. Generalized effect size.}

\item{alpha}{numeric value. Probability (range 0.01-1.00)}

\item{power}{numeric value. Statistical power}

\item{lost}{numeric value. Expected proportion of non-informative pairs.}

\item{verbose}{logical value. If TRUE, results output is in text format.}
}
\description{
Study sample size calculation using the generalized effect size method
}
