% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplanMeier_PFSr.R
\name{kaplanMeier_PFSr}
\alias{kaplanMeier_PFSr}
\title{Analysis of progression free survival ratio (PFSr) using the Kaplan-Meier method}
\usage{
kaplanMeier_PFSr(data, delta = 1, plot = FALSE)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric)}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}

\item{plot}{a logical value. Indication if a plot should be generated.}
}
\value{
Kaplan-Meier based statistics of PFSr and optionally a Kaplan-Meier plot
}
\description{
Analysis of progression free survival ratio (PFSr) using the Kaplan-Meier method
}
\examples{
data(input)
kaplanMeier_PFSr(data = data, delta = 1.3)
}
