% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelKM_PFSr.R
\name{kernelKM_PFSr}
\alias{kernelKM_PFSr}
\title{Kernel conditional Kaplan Meier method}
\usage{
kernelKM_PFSr(
  data,
  ratio.name,
  status.name,
  time0.name,
  points = NULL,
  conf.int = FALSE,
  n.boot = 2000
)
}
\arguments{
\item{data}{a data frame with values for PFS1/PFS2 ratio, status (1/0) and PFS1 (time0)}

\item{ratio.name}{name of the column that contains the PFS1/PFS2 ratio ('numeric')}

\item{status.name}{name of the column that contains the status ('boolean')}

\item{time0.name}{name of the column that contains the PFS1 ('numeric')}

\item{points}{}

\item{conf.int}{a boolean indicating if confidende intervals should be calculated}

\item{n.boot}{number of permutations}
}
\value{
a Kaplan-Meier based statistics according to the kernel conditional KM method
}
\description{
Kernel conditional Kaplan Meier method
}
\examples{
# Example with default values:
kernelKM_PFSr(data, "ratio", "status", "PFS1")
# Example with modified values
kernelKM_PFSr(data, "ratio", "status", "PFS1", points = 2, conf.int = TRUE)
}
