% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midrank_PFSr.R
\name{midrank_PFSr}
\alias{midrank_PFSr}
\title{Analysis of progression free survival ratio (PFSr) using the midrank method}
\usage{
midrank_PFSr(data, delta = 1)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric)}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}
}
\value{
midrank-based statistics for PFSr analysis
}
\description{
Analysis of progression free survival ratio (PFSr) using the midrank method
}
\examples{
data(input)
midrank_PFSr(data = input, delta = 1.3)
}
