% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_PFS.R
\name{modify_PFS}
\alias{modify_PFS}
\title{Modified progression free survival ratio (mPFSr)}
\usage{
modify_PFS(data, delta = 1, min_pfs2 = 6)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric)}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}

\item{min_pfs2}{a numeric value. The value 'min_pfs2' is the minimum PFS2 in month that can be considered satisfying irrespective of PFS1.}
}
\value{
dataframe with modified PFS and PFSr.
}
\description{
Modified progression free survival ratio (mPFSr)
}
\examples{
data(input)
modify_PFS(data = input, delta = 1.3, min_pfs2 = 6)
}
