% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_PFSr.R
\name{parametric_PFSr}
\alias{parametric_PFSr}
\title{Analysis of progression free survival ratio (PFSr) using the parametric method}
\usage{
parametric_PFSr(data, delta = 1)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) ) and 'ratio' (numeric, ratio PFS2/PFS1)}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}
}
\value{
Parametric-based statistics for PFSr analysis
}
\description{
Analysis of progression free survival ratio (PFSr) using the parametric method
}
\examples{
data(input)
parametric_PFSr(data = data, delta = 1.3)
}
