% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSr_power_calculation_ges.R
\name{PFSr_power_calculation_ges}
\alias{PFSr_power_calculation_ges}
\title{Study power calculation using the GES method}
\usage{
PFSr_power_calculation_ges(
  sample_size = NULL,
  ges = 0.3,
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{sample_size}{a numeric value. The sample size of the study.}

\item{ges}{a numeric value. Generalized effect size.}

\item{alpha}{a numeric value. Probability (range 0.01-1.00)}

\item{verbose}{logical value. If TRUE, results output is in text format.}
}
\value{
the calculated required sample size for the defined parameters
}
\description{
Study power calculation using the GES method
}
