% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_PFSr.R
\name{PFSr_samplesize}
\alias{PFSr_samplesize}
\title{Study sample size calculation}
\usage{
PFSr_samplesize(
  alpha = 0.05,
  power = 0.8,
  rho = 0.5,
  alt_HR = 1.33,
  null_HR = 1,
  k = 1,
  lost = 0.1,
  model = c("GBVE", "Weibull"),
  verbose = FALSE
)
}
\arguments{
\item{alpha}{a numeric value. between 0.01-1.00. Probability}

\item{power}{a numeric value. statistical power}

\item{rho}{a numeric value. the correlation between PFS1 and PFS2.}

\item{alt_HR}{a numeric value. H~1~ (satisfying PFSr).}

\item{null_HR}{a numeric value. H~0~ (null hypothesis, unsatisfying PFSr).}

\item{k}{a numeric value. Common shape parameter k (k </=/> 1 for decreasing, constant, and increasing hazard functions).}

\item{lost}{a numeric value. expected proportion of non-informative pairs}

\item{model}{a factor with levels c("GBVE", "Weibull"). Indicates the model used for the calculation.}

\item{verbose}{if TRUE, results output is in text format}
}
\value{
a dataframe with the calculated required sample size for the
}
\description{
Study sample size calculation
}
