% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_PFSr_proportion.R
\name{PFSr_samplesize_proportion}
\alias{PFSr_samplesize_proportion}
\title{Study sample size calculation according to a pre-specified proportion of patients with positive PFSr}
\usage{
PFSr_samplesize_proportion(
  p0,
  p1,
  alpha,
  beta,
  pfsratio,
  lost = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{p0}{a numeric value. Proportion of patients with a PFSratio > pfsratio according to the null hypothesis}

\item{p1}{a numeric value. True proportion of patients with a PFSratio > pfsratio}

\item{alpha}{a numeric value. Probability (range 0.01-1.00)}

\item{beta}{a numeric value.}

\item{pfsratio}{a numeric value. Desired PFSr that the study should achieve.}

\item{lost}{a numeric value. Expected proportion of non-informative pairs.}

\item{verbose}{a logical value. If TRUE, results output is in text format.}
}
\description{
Study sample size calculation according to a pre-specified proportion of patients with positive PFSr
}
