% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_PFS.R
\name{plot_correlation_PFS}
\alias{plot_correlation_PFS}
\title{Plot the correlation of PFS values}
\usage{
plot_correlation_PFS(data, delta = 1, log_scale = FALSE)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) )}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}

\item{log_scale}{a logical value indicating if the scale should be log transformed.}
}
\value{
A plot of the correlation between PFS1 and PFS2.
}
\description{
Plot the correlation of PFS values
}
\examples{
data(input)
plot_correlation_PFS(data = data, delta = 1.3, log_scale = TRUE)
}
