% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumHaz.R
\name{plot_cumHaz}
\alias{plot_cumHaz}
\title{Plot cumulative hazard ratio}
\usage{
plot_cumHaz(data, selected_PFS = c("PFS1", "PFS2"))
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) )}

\item{selected_PFS}{a vector of selected PFS values, by default c("PFS1", "PFS2")}
}
\value{
A plot of the cumulative hazard ratio.
}
\description{
Plot cumulative hazard ratio
}
\examples{
data(input)
plot_cumHaz(data = input, selected_PFS = "PFS2")
}
