% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kernelKM_PFSr.R
\name{plot_kernelKM_PFSr}
\alias{plot_kernelKM_PFSr}
\title{Plot results from the kernel conditional Kaplan-Meier statistics}
\usage{
plot_kernelKM_PFSr(res_kernelKM_PFSr)
}
\arguments{
\item{res_kernelKM_PFSr}{results from the kernelKM_PFSr function, min requirement is a list with elements ratio and surv}
}
\value{
a ggplot object displaying a Kaplan-Meier curve for survival data calculated using the kernel conditional KM method
}
\description{
Plot results from the kernel conditional Kaplan-Meier statistics
}
\examples{
res <- kernelKM_PFSr(data = input, ratio.name = "ratio", status.name = "status", time0.name = "PFS1", conf.int = TRUE, points = 2)
plot_kernelKM_PFSr(res_kernelKM_PFSr = res)
}
