% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_PFS.R
\name{summarize_PFS}
\alias{summarize_PFS}
\title{Summarise statistics of the PFS values}
\usage{
summarize_PFS(data)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) ) and 'ratio' (numeric, ratio PFS2/PFS1)}
}
\value{
A table with summary statistics on the PFS values.
}
\description{
Summarise statistics of the PFS values
}
\examples{
data(input)
input$ratio <- input$PFS2/input$PFS1
summarize_PFS(data = input)
}
