% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmerplot_PFSr.R
\name{swimmerplot_PFSr}
\alias{swimmerplot_PFSr}
\title{Swimmer plot for individual PFS values}
\usage{
swimmerplot_PFSr(data, delta = 1)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) )}

\item{delta}{a numeric value. The value 'delta' is the desired difference between PFS2 and PFS1 that is seen as a success.}
}
\value{
A plot of individual PFS1 and PFS2 values, sorted by the PFSr.
}
\description{
Swimmer plot for individual PFS values
}
\examples{
data(input)
input$ratio <- input$PFS2/input$PFS1
swimmerplot_PFSr(data = input, delta = 1.3)
}
