% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_weibull_distribution.R
\name{test_weibull_distribution}
\alias{test_weibull_distribution}
\title{Test if the PFS values follow a Weibull distribution}
\usage{
test_weibull_distribution(data, tidy = TRUE)
}
\arguments{
\item{data}{a dataframe with columns 'PFS1' (numeric), 'PF2' (numeric), 'status' (logical or integer with values c(0,1) )}

\item{tidy}{a logical value indicating if the data should be returned in the tidy data format}
}
\value{
a table with Weibull survival statistics
}
\description{
Test if the PFS values follow a Weibull distribution
}
\examples{
data(input)
test_weibull_distribution(data = input, tidy = TRUE)
}
